LayoutTests/http/wpt/resource-timing/rt-nextHopProtocol-expected.txt

Resource Timing: nextHopProtocol


PASS nextHopProtocol is expected to be 'http1/1'
